+++ dart/io/urdf/urdf_parser.cpp
@@ -67,107 +67,6 @@
 namespace dart {
 namespace io {
 
-using ModelInterfacePtr = std::shared_ptr<urdf::ModelInterface>;
-
-//==============================================================================
-UrdfParser::Options::Options(
-    common::ResourceRetrieverPtr resourceRetriever,
-    RootJointType defaultRootJointType,
-    const dynamics::Inertia& defaultInertia)
-  : mResourceRetriever(std::move(resourceRetriever)),
-    mDefaultRootJointType(defaultRootJointType),
-    mDefaultInertia(defaultInertia)
-{
-  // Do nothing
-}
-
-//==============================================================================
-UrdfParser::UrdfParser(const Options& options)
-  : mOptions(options),
-    mLocalRetriever(new common::LocalResourceRetriever),
-    mPackageRetriever(new io::PackageResourceRetriever(mLocalRetriever)),
-    mRetriever(new io::CompositeResourceRetriever)
-{
-  mRetriever->addSchemaRetriever("file", mLocalRetriever);
-  mRetriever->addSchemaRetriever("package", mPackageRetriever);
-  mRetriever->addSchemaRetriever("dart", DartResourceRetriever::create());
-}
-
-//==============================================================================
-UrdfParser::~UrdfParser() = default;
-
-//==============================================================================
-void UrdfParser::setOptions(const Options& options)
-{
-  mOptions = options;
-}
-
-//==============================================================================
-const UrdfParser::Options& UrdfParser::getOptions() const
-{
-  return mOptions;
-}
-
-//==============================================================================
-void UrdfParser::addPackageDirectory(
-    std::string_view packageName, std::string_view packageDirectory)
-{
-  mPackageRetriever->addPackageDirectory(packageName, packageDirectory);
-}
-
-//==============================================================================
-dynamics::SkeletonPtr UrdfParser::parseSkeleton(const common::Uri& uri)
-{
-  const common::ResourceRetrieverPtr resourceRetriever
-      = getResourceRetriever(mOptions.mResourceRetriever);
-
-  std::string content;
-  if (!readFileToString(resourceRetriever, uri, content)) {
-    return nullptr;
-  }
-
-  // Use urdfdom to load the URDF file.
-  const ModelInterfacePtr urdfInterface = urdf::parseURDF(content);
-  if (!urdfInterface) {
-    DART_WARN("Failed loading URDF file '{}'.", uri.toString());
-    return nullptr;
-  }
-
-  ParseContext context;
-  context.mTransmissions = parseTransmissions(content);
-
-  return modelInterfaceToSkeleton(
-      urdfInterface.get(), uri, resourceRetriever, mOptions, &context);
-}
-
-//==============================================================================
-dynamics::SkeletonPtr UrdfParser::parseSkeletonString(
-    std::string_view urdfString, const common::Uri& baseUri)
-{
-  if (urdfString.empty()) {
-    DART_WARN(
-        "A blank string cannot be parsed into a Skeleton. Returning a nullptr");
-    return nullptr;
-  }
-
-  const std::string urdfStringCopy(urdfString);
-  ModelInterfacePtr urdfInterface = urdf::parseURDF(urdfStringCopy);
-  if (!urdfInterface) {
-    DART_WARN("Failed loading URDF.");
-    return nullptr;
-  }
-
-  ParseContext context;
-  context.mTransmissions = parseTransmissions(urdfStringCopy);
-
-  return modelInterfaceToSkeleton(
-      urdfInterface.get(),
-      baseUri,
-      getResourceRetriever(mOptions.mResourceRetriever),
-      mOptions,
-      &context);
-}
-
 dynamics::SkeletonPtr UrdfParser::modelInterfaceToSkeleton(
     const urdf::ModelInterface* model,
     const common::Uri& baseUri,
